# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Library Management",
    "version": "2.0.0",
    "description": "A comprehensive API for managing library books and authors with authentication",
    "contact": {
      "name": "Library API Support",
      "email": "support@libraryapi.com"
    }
  },
  "servers": [
    {
      "url": "https://library-management-d0no.onrender.com",
      "description": "Production server (Render)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
...
      "Author": {
        "type": "object",
        "required": ["name", "email"],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Auto-generated ID"
          },
          "name": {
            "type": "string",
            "description": "Author's full name",
            "example": "J.K. Rowling"
          },
          "email": {
            "type": "string",
            "description": "Author's email address",
            "example": "jk.rowling@example.com"
          },
          "biography": {
            "type": "string",
            "description": "Author's biography",
            "example": "British author best known for the Harry Potter series"
          },
          "birthDate": {
            "type": "string",
            "format": "date",
            "description": "Author's birth date",
            "example": "1965-07-31"
          },
          "nationality": {
            "type": "string",
            "description": "Author's nationality",
            "example": "British"
          }
        }
      },
      "Book": {
        "type": "object",
        "required": ["title", "authorId", "isbn"],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Auto-generated ID"
          },
          "title": {
            "type": "string",
            "description": "Book title",
            "example": "Harry Potter and the Philosopher's Stone"
          },
          "authorId": {
            "type": "string",
            "description": "Author's ID",
            "example": "68853c17591d38453d07889e"
          },
          "isbn": {
            "type": "string",
            "description": "ISBN number",
            "example": "978-0-7475-3269-9"
          },
          "publishedYear": {
            "type": "number",
            "description": "Year of publication",
            "example": 1997
          },
          "genre": {
            "type": "string",
            "description": "Book genre",
            "example": "Fantasy"
          },
          "pages": {
            "type": "number",
            "description": "Number of pages",
            "example": 223
          },
          "publisher": {
            "type": "string",
            "description": "Publisher name",
            "example": "Bloomsbury"
          },
          "language": {
            "type": "string",
            "description": "Book language",
            "example": "English"
          },
          "description": {
            "type": "string",
            "description": "Book description",
            "example": "A young wizard discovers his magical heritage and begins his education at Hogwarts School of Witchcraft and Wizardry"
          },
          "price": {
            "type": "number",
            "description": "Book price",
            "example": 14.99
          },
          "inStock": {
            "type": "boolean",
            "description": "Availability status",
            "example": true
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "message": {
            "type": "string",
            "description": "Detailed error description"
          },
          "details": {
            "type": "string",
            "description": "Validation error details"
          }
        }
      },
...
    },
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "library.sid",
        "description": "Session-based authentication using cookies"
      }
    }
  },
...
  "paths": {
...
    "/authors": {
      "get": {
        "summary": "Get all authors",
        "tags": ["Authors"],
        "responses": {
          "200": {
            "description": "List of all authors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Author"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new author (Protected)",
        "tags": ["Authors"],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Author created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Author created successfully"
                    },
                    "authorId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/books": {
      "get": {
        "summary": "Get all books",
        "tags": ["Books"],
        "responses": {
          "200": {
            "description": "List of all books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new book (Protected)",
        "tags": ["Books"],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Book created successfully"
                    },
                    "bookId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}